added hs account command, moved rename and list to it, and removed them from hs config command
