+++ packages/cli/bin/cli.js
@@ -31,6 +31,7 @@ const mvCommand = require('../commands/mv');
 const projectsCommand = require('../commands/projects');
 const themeCommand = require('../commands/theme');
 const configCommand = require('../commands/config');
+const accountsCommand = require('../commands/accounts');
 
 const notifier = updateNotifier({ pkg: { ...pkg, name: '@hubspot/cli' } });
 
@@ -88,6 +89,7 @@ const argv = yargs
   .command(projectsCommand)
   .command(themeCommand)
   .command(configCommand)
+  .command(accountsCommand)
   .help()
   .recommendCommands()
   .demandCommand(1, '')
