+++ packages/cli/commands/accounts/rename.js
@@ -1,14 +1,10 @@
 const { logger } = require('@hubspot/cli-lib/logger');
-const { getConfig, getConfigPath } = require('@hubspot/cli-lib/lib/config');
+const { renameAccount } = require('@hubspot/cli-lib/lib/config');
 const {
   loadConfig,
   validateConfig,
   checkAndWarnGitInclusion,
 } = require('@hubspot/cli-lib');
-const {
-  getTableContents,
-  getTableHeader,
-} = require('@hubspot/cli-lib/lib/table');
 
 const {
   addConfigOptions,
@@ -32,34 +28,36 @@ const loadAndValidateOptions = async options => {
   }
 };
 
-exports.command = 'list';
-exports.describe = 'List names of accounts defined in config';
+exports.command = 'rename <accountName> <newName>';
+exports.describe = 'Rename account in config';
 
 exports.handler = async options => {
   loadAndValidateOptions(options);
 
+  const { accountName, newName } = options;
   const accountId = getAccountId(options);
 
-  trackCommandUsage('config-list', {}, accountId);
+  trackCommandUsage('accounts-rename', {}, accountId);
 
-  const config = getConfig();
-  const configPath = getConfigPath();
-  const portalData = config.portals.map(portal => {
-    return [portal.name, portal.portalId, portal.authType];
-  });
-  portalData.unshift(getTableHeader(['Name', 'Account ID', 'Auth Type']));
+  await renameAccount(accountName, newName);
 
-  logger.log(`Config path: ${configPath}`);
-  logger.log('Default account: ', config.defaultPortal);
-  logger.log('Accounts:');
-  logger.log(getTableContents(portalData, { border: { bodyLeft: '  ' } }));
+  return logger.log(`Account ${accountName} renamed to ${newName}`);
 };
 
 exports.builder = yargs => {
   addConfigOptions(yargs, true);
   addAccountOptions(yargs, true);
 
-  yargs.example([['$0 config list']]);
+  yargs.positional('accountName', {
+    describe: 'Name of account to be renamed.',
+    type: 'string',
+  });
+  yargs.positional('newName', {
+    describe: 'New name for account.',
+    type: 'string',
+  });
+
+  yargs.example([['$0 accounts rename myExistingPortalName myNewPortalName']]);
 
   return yargs;
 };
