+++ packages/cli/commands/config.js
@@ -1,7 +1,5 @@
 const { addConfigOptions, addAccountOptions } = require('../lib/commonOpts');
-const list = require('./config/list');
 const set = require('./config/set');
-const rename = require('./config/rename');
 
 exports.command = 'config';
 exports.describe = 'Commands for working with the config file';
@@ -10,14 +8,7 @@ exports.builder = yargs => {
   addConfigOptions(yargs, true);
   addAccountOptions(yargs, true);
 
-  yargs
-    .command({
-      ...list,
-      aliases: 'ls',
-    })
-    .command(set)
-    .command(rename)
-    .demandCommand(1, '');
+  yargs.command(set).demandCommand(1, '');
 
   return yargs;
 };
