+++ packages/cli/commands/config/set.js
@@ -1,5 +1,5 @@
 const { addConfigOptions, addAccountOptions } = require('../../lib/commonOpts');
-const defaultCommand = require('./set/default');
+const defaultAccount = require('./set/default-account');
 
 exports.command = 'set';
 exports.describe = 'Commands for working with the config file';
@@ -8,7 +8,7 @@ exports.builder = yargs => {
   addConfigOptions(yargs, true);
   addAccountOptions(yargs, true);
 
-  yargs.command(defaultCommand).demandCommand(1, '');
+  yargs.command(defaultAccount).demandCommand(1, '');
 
   return yargs;
 };
